+++ src/socketlib/services/abstract_service.py
@@ -2,7 +2,7 @@ import abc
 import logging
 import queue
 import threading
-from typing import Any, Callable, Optional
+from typing import Callable, Optional
 
 
 class AbstractService(abc.ABC):
@@ -26,6 +26,14 @@ class AbstractService(abc.ABC):
             events: Optional[dict[str, threading.Event]] = None,
             logger: Optional[logging.Logger] = None,
     ):
+        """Initialize the AbstractService.
+
+            :param in_queue: Input queue for receiving messages.
+            :param out_queue: Output queue for sending messages.
+            :param stop: Custom stop function.
+            :param events: Dictionary of events.
+            :param logger: Logger for logging events.
+        """
         self._in = in_queue if in_queue is not None else queue.Queue()
         self._out = out_queue if out_queue is not None else queue.Queue()
         self._events = events if events is not None else dict()
@@ -40,10 +48,12 @@ class AbstractService(abc.ABC):
 
     @property
     def process_thread(self) -> threading.Thread:
+        """The thread responsible for processing messages."""
         return self._process_thread
 
     @property
     def in_queue(self) -> queue.Queue:
+        """The input queue for receiving messages."""
         return self._in
 
     @property
@@ -52,6 +62,7 @@ class AbstractService(abc.ABC):
 
     @abc.abstractmethod
     def _handle_message(self):
+        """Abstract method to handle incoming messages."""
         raise NotImplementedError
 
     def start_main_thread(self):
@@ -59,11 +70,13 @@ class AbstractService(abc.ABC):
         self._handle_message()
 
     def start(self) -> None:
+        """Starts the service by launching the processing thread."""
         self._process_thread.start()
         if self._logger is not None:
             self._logger.debug(f"Started {self.__class__.__name__}")
 
     def join(self) -> None:
+        """Waits for the processing thread to finish."""
         self._process_thread.join()
 
     @staticmethod
@@ -71,11 +84,13 @@ class AbstractService(abc.ABC):
             stop: Optional[Callable[[], bool]],
             stop_event: threading.Event
     ) -> Callable[[], bool]:
+        """Returns a stop function based on the provided arguments."""
         if stop is None:
             return lambda: stop_event.is_set()
         return stop
 
     def shutdown(self) -> None:
+        """Shuts down the service gracefully."""
         self._stop_event.set()
         self.join()
         if self._logger is not None: